+++ app/controllers/api/account.php
@@ -2634,6 +2634,10 @@ App::post('/v1/account/verification')
             throw new Exception(Exception::GENERAL_SMTP_DISABLED, 'SMTP Disabled');
         }
 
+        if ($user->getAttribute('emailVerification')) {
+            throw new Exception(Exception::USER_EMAIL_ALREADY_VERIFIED);
+        }
+
         $roles = Authorization::getRoles();
         $isPrivilegedUser = Auth::isPrivilegedUser($roles);
         $isAppUser = Auth::isAppUser($roles);
@@ -2851,6 +2855,10 @@ App::post('/v1/account/verification/phone')
 
         if (empty($user->getAttribute('phone'))) {
             throw new Exception(Exception::USER_PHONE_NOT_FOUND);
+        }
+
+        if ($user->getAttribute('phoneVerification')) {
+            throw new Exception(Exception::USER_PHONE_ALREADY_VERIFIED);
         }
 
         $roles = Authorization::getRoles();